requirements.txt
aiofiles==0.5.0
aniso8601==7.0.0
appdirs==1.4.3
async-exit-stack==1.0.1
async-generator==1.10
bcrypt==3.2.0
CacheControl==0.12.6
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dnspython==2.1.0
ecdsa==0.14.1
email-validator==1.1.2
fastapi==0.63.0
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
h11==0.12.0
html5lib==1.0.1
httptools==0.1.1
idna==2.8
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
MarkupSafe==1.1.1
msgpack==0.6.2
orjson==3.4.7
packaging==20.3
passlib==1.7.4
pep517==0.8.2
progress==1.5
promise==2.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.3
pymongo==3.11.2
pyparsing==2.4.6
python-dotenv==0.15.0
python-jose==3.2.0
python-multipart==0.0.5
pytoml==0.1.21
PyYAML==5.4.1
requests==2.25.1
retrying==1.3.3
rsa==4.7
Rx==1.6.1
six==1.14.0
starlette==0.13.6
ujson==3.2.0
urllib3==1.25.8
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
webencodings==0.5.1
websockets==8.1
